requirements.txt
alembic==1.8.0
black==22.3.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
mariadb==1.0.11
MarkupSafe==2.1.1
marshmallow==3.16.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
python-decouple==3.6
six==1.16.0
SQLAlchemy==1.4.37
tomli==2.0.1
Werkzeug==2.1.2
